src/azure-function-unzip-large-files/function-unzip-large-zip-files/requirements.txt
azure-functions
azure-storage-blob
pyzipper
python-dotenv